# openapi-eventia.yaml
openapi: 3.0.1
info:
  title: Eventia API
  description: API REST para la gestión de eventos, artistas, recintos, usuarios y reservas.
  version: 1.0.0
servers:
  - url: http://localhost:8080
    description: Servidor local
...
paths:
...
  /eventos:
    get:
      tags: [Eventos]
      summary: Obtener todos los eventos
      description: Filtra opcionalmente por nombre, categoria y cancelado.
      parameters:
        - name: nombre
          in: query
          required: false
          description: Filtra por nombre del evento.
          schema:
            type: string
          example: Festival de Primavera
        - name: categoria
          in: query
          required: false
          description: Filtra por categoría del evento.
          schema:
            type: string
          example: Festival
        - name: cancelado
          in: query
          required: false
          description: Filtra por si el evento está cancelado.
          schema:
            type: boolean
          example: false
      responses:
        '200':
          description: Lista de eventos obtenida correctamente
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Evento'
              examples:
                sinFiltros:
                  summary: Consulta sin filtros
                  value:
                    - id: 1
                      nombre: Festival de Primavera
                      descripcion: Festival de música en directo
                      fechaEvento: '2026-06-20'
                      horaEvento: '21:00:00'
                      precioEntrada: 45.5
                      aforoMaximo: 12000
                      entradasDisponibles: 8500
                      cancelado: false
                      presencial: true
                      categoria: Festival
                      usuario:
                        id: 1
                      artista:
                        id: 1
                      recinto:
                        id: 1
                    - id: 2
                      nombre: Concierto Acústico
                      descripcion: Noche íntima de música acústica
                      fechaEvento: '2026-07-15'
                      horaEvento: '20:00:00'
                      precioEntrada: 25.0
                      aforoMaximo: 500
                      entradasDisponibles: 200
                      cancelado: false
                      presencial: true
                      categoria: Concierto
                      usuario:
                        id: 2
                      artista:
                        id: 2
                      recinto:
                        id: 2
                filtradoPorCategoria:
                  summary: Consulta filtrada por categoría Festival
                  value:
                    - id: 1
                      nombre: Festival de Primavera
                      descripcion: Festival de música en directo
                      fechaEvento: '2026-06-20'
                      horaEvento: '21:00:00'
                      precioEntrada: 45.5
                      aforoMaximo: 12000
                      entradasDisponibles: 8500
                      cancelado: false
                      presencial: true
                      categoria: Festival
                      usuario:
                        id: 1
                      artista:
                        id: 1
                      recinto:
                        id: 1
        '400':
          $ref: '#/components/responses/BadRequest'
        '500':
          $ref: '#/components/responses/InternalServerError'

    post:
      tags: [Eventos]
      summary: Crear evento
      description: Registra un nuevo evento.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Evento'
            examples:
              ejemplo1:
                summary: Festival
                value:
                  nombre: Festival de Primavera
                  descripcion: Festival de música en directo
                  fechaEvento: '2026-06-20'
                  horaEvento: '21:00:00'
                  precioEntrada: 45.5
                  aforoMaximo: 12000
                  entradasDisponibles: 12000
                  cancelado: false
                  presencial: true
                  categoria: Festival
                  usuario:
                    id: 1
                  artista:
                    id: 1
                  recinto:
                    id: 1
              ejemplo2:
                summary: Concierto
                value:
                  nombre: Concierto Acústico
                  descripcion: Noche íntima de música acústica
                  fechaEvento: '2026-07-15'
                  horaEvento: '20:00:00'
                  precioEntrada: 25.0
                  aforoMaximo: 500
                  entradasDisponibles: 500
                  cancelado: false
                  presencial: true
                  categoria: Concierto
                  usuario:
                    id: 2
                  artista:
                    id: 2
                  recinto:
                    id: 2
      responses:
        '201':
          description: Evento creado correctamente
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Evento'
              examples:
                ejemplo1:
                  summary: Festival creado
                  value:
                    id: 1
                    nombre: Festival de Primavera
                    descripcion: Festival de música en directo
                    fechaEvento: '2026-06-20'
                    horaEvento: '21:00:00'
                    precioEntrada: 45.5
                    aforoMaximo: 12000
                    entradasDisponibles: 12000
                    cancelado: false
                    presencial: true
                    categoria: Festival
                    usuario:
                      id: 1
                    artista:
                      id: 1
                    recinto:
                      id: 1
                ejemplo2:
                  summary: Concierto creado
                  value:
                    id: 2
                    nombre: Concierto Acústico
                    descripcion: Noche íntima de música acústica
                    fechaEvento: '2026-07-15'
                    horaEvento: '20:00:00'
                    precioEntrada: 25.0
                    aforoMaximo: 500
                    entradasDisponibles: 500
                    cancelado: false
                    presencial: true
                    categoria: Concierto
                    usuario:
                      id: 2
                    artista:
                      id: 2
                    recinto:
                      id: 2
        '400':
          $ref: '#/components/responses/BadRequest'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /eventos/{id}:
    get:
      tags: [Eventos]
      summary: Obtener evento por id
      parameters:
        - $ref: '#/components/parameters/IdPath'
      responses:
        '200':
          description: Evento obtenido correctamente
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Evento'
              examples:
                ejemplo1:
                  summary: Festival
                  value:
                    id: 1
                    nombre: Festival de Primavera
                    descripcion: Festival de música en directo
                    fechaEvento: '2026-06-20'
                    horaEvento: '21:00:00'
                    precioEntrada: 45.5
                    aforoMaximo: 12000
                    entradasDisponibles: 8500
                    cancelado: false
                    presencial: true
                    categoria: Festival
                    usuario:
                      id: 1
                    artista:
                      id: 1
                    recinto:
                      id: 1
                ejemplo2:
                  summary: Concierto
                  value:
                    id: 2
                    nombre: Concierto Acústico
                    descripcion: Noche íntima de música acústica
                    fechaEvento: '2026-07-15'
                    horaEvento: '20:00:00'
                    precioEntrada: 25.0
                    aforoMaximo: 500
                    entradasDisponibles: 200
                    cancelado: false
                    presencial: true
                    categoria: Concierto
                    usuario:
                      id: 2
                    artista:
                      id: 2
                    recinto:
                      id: 2
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'

    put:
      tags: [Eventos]
      summary: Modificar evento
      description: Sustituye completamente los datos de un evento existente.
      parameters:
        - $ref: '#/components/parameters/IdPath'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Evento'
            examples:
              ejemplo1:
                summary: Actualizar precio
                value:
                  nombre: Festival de Primavera
                  descripcion: Festival de música en directo con nuevo cartel
                  fechaEvento: '2026-06-20'
                  horaEvento: '21:00:00'
                  precioEntrada: 55.0
                  aforoMaximo: 12000
                  entradasDisponibles: 8500
                  cancelado: false
                  presencial: true
                  categoria: Festival
                  usuario:
                    id: 1
                  artista:
                    id: 1
                  recinto:
                    id: 1
              ejemplo2:
                summary: Actualizar hora
                value:
                  nombre: Concierto Acústico
                  descripcion: Noche íntima de música acústica
                  fechaEvento: '2026-07-15'
                  horaEvento: '19:00:00'
                  precioEntrada: 25.0
                  aforoMaximo: 500
                  entradasDisponibles: 200
                  cancelado: false
                  presencial: true
                  categoria: Concierto
                  usuario:
                    id: 2
                  artista:
                    id: 2
                  recinto:
                    id: 2
      responses:
        '200':
          description: Evento modificado correctamente
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Evento'
              examples:
                ejemplo1:
                  summary: Evento con precio actualizado
                  value:
                    id: 1
                    nombre: Festival de Primavera
                    descripcion: Festival de música en directo con nuevo cartel
                    fechaEvento: '2026-06-20'
                    horaEvento: '21:00:00'
                    precioEntrada: 55.0
                    aforoMaximo: 12000
                    entradasDisponibles: 8500
                    cancelado: false
                    presencial: true
                    categoria: Festival
                    usuario:
                      id: 1
                    artista:
                      id: 1
                    recinto:
                      id: 1
                ejemplo2:
                  summary: Evento con hora actualizada
                  value:
                    id: 2
                    nombre: Concierto Acústico
                    descripcion: Noche íntima de música acústica
                    fechaEvento: '2026-07-15'
                    horaEvento: '19:00:00'
                    precioEntrada: 25.0
                    aforoMaximo: 500
                    entradasDisponibles: 200
                    cancelado: false
                    presencial: true
                    categoria: Concierto
                    usuario:
                      id: 2
                    artista:
                      id: 2
                    recinto:
                      id: 2
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'

    patch:
      tags: [Eventos]
      summary: Modificar parcialmente un evento
      description: Actualiza uno o varios campos de un evento existente.
      parameters:
        - $ref: '#/components/parameters/IdPath'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: true
            examples:
              ejemplo1:
                summary: Cancelar evento
                value:
                  cancelado: true
              ejemplo2:
                summary: Actualizar precio y entradas disponibles
                value:
                  precioEntrada: 39.9
                  entradasDisponibles: 7000
      responses:
        '200':
          description: Evento actualizado parcialmente
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Evento'
              examples:
                ejemplo1:
                  summary: Evento cancelado
                  value:
                    id: 1
                    nombre: Festival de Primavera
                    descripcion: Festival de música en directo
                    fechaEvento: '2026-06-20'
                    horaEvento: '21:00:00'
                    precioEntrada: 45.5
                    aforoMaximo: 12000
                    entradasDisponibles: 8500
                    cancelado: true
                    presencial: true
                    categoria: Festival
                    usuario:
                      id: 1
                    artista:
                      id: 1
                    recinto:
                      id: 1
                ejemplo2:
                  summary: Evento con precio actualizado
                  value:
                    id: 1
                    nombre: Festival de Primavera
                    descripcion: Festival de música en directo
                    fechaEvento: '2026-06-20'
                    horaEvento: '21:00:00'
                    precioEntrada: 39.9
                    aforoMaximo: 12000
                    entradasDisponibles: 7000
                    cancelado: false
                    presencial: true
                    categoria: Festival
                    usuario:
                      id: 1
                    artista:
                      id: 1
                    recinto:
                      id: 1
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'

    delete:
      tags: [Eventos]
      summary: Eliminar evento
      parameters:
        - $ref: '#/components/parameters/IdPath'
      responses:
        '204':
          description: Evento eliminado correctamente
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /eventos/cancelados:
    get:
      tags: [Eventos]
      summary: Obtener eventos cancelados
      description: Devuelve los eventos con cancelado = true mediante una consulta JPQL.
      responses:
        '200':
          description: Lista de eventos cancelados obtenida correctamente
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Evento'
              examples:
                ejemplo1:
                  summary: Un evento cancelado
                  value:
                    - id: 3
                      nombre: Macrofestival Verano
                      descripcion: Cancelado por causas meteorológicas
                      fechaEvento: '2026-08-01'
                      horaEvento: '18:00:00'
                      precioEntrada: 60.0
                      aforoMaximo: 50000
                      entradasDisponibles: 30000
                      cancelado: true
                      presencial: true
                      categoria: Festival
                      usuario:
                        id: 1
                      artista:
                        id: 3
                      recinto:
                        id: 2
                ejemplo2:
                  summary: Varios eventos cancelados
                  value:
                    - id: 3
                      nombre: Macrofestival Verano
                      descripcion: Cancelado por causas meteorológicas
                      fechaEvento: '2026-08-01'
                      horaEvento: '18:00:00'
                      precioEntrada: 60.0
                      aforoMaximo: 50000
                      entradasDisponibles: 30000
                      cancelado: true
                      presencial: true
                      categoria: Festival
                      usuario:
                        id: 1
                      artista:
                        id: 3
                      recinto:
                        id: 2
                    - id: 5
                      nombre: Concierto Privado
                      descripcion: Evento privado cancelado
                      fechaEvento: '2026-09-10'
                      horaEvento: '22:00:00'
                      precioEntrada: 80.0
                      aforoMaximo: 200
                      entradasDisponibles: 50
                      cancelado: true
                      presencial: true
                      categoria: Privado
                      usuario:
                        id: 2
                      artista:
                        id: 1
                      recinto:
                        id: 1
        '500':
          $ref: '#/components/responses/InternalServerError'
...
components:
  parameters:
    IdPath:
      name: id
      in: path
      required: true
      description: Identificador único del recurso
      schema:
        type: integer
        format: int64
        minimum: 1
      example: 1
  responses:
    BadRequest:
      description: Error de validación o petición incorrecta
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            code: 400
            title: bad-request
            message: Error de validación
            errors:
              email: El email no es válido
              password: La contraseña debe tener entre 6 y 20 caracteres
    NotFound:
      description: Recurso no encontrado
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            code: 404
            title: not-found
            message: Recurso no encontrado
            errors: {}
    InternalServerError:
      description: Error interno del servidor
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            code: 500
            title: internal-server-error
            message: Error interno del servidor
            errors: {}
  schemas:
    ErrorResponse:
      type: object
      properties:
        code:
          type: integer
          example: 400
        title:
          type: string
          example: bad-request
        message:
          type: string
          example: Error de validación
        errors:
          type: object
          additionalProperties:
            type: string
      required:
        - code
        - title
        - message
        - errors
    UsuarioRef:
      type: object
      properties:
        id:
          type: integer
          format: int64
          example: 1
      required:
        - id
    ArtistaRef:
      type: object
      properties:
        id:
          type: integer
          format: int64
          example: 1
      required:
        - id
    RecintoRef:
      type: object
      properties:
        id:
          type: integer
          format: int64
          example: 1
      required:
        - id
...
    Evento:
      type: object
      properties:
        id:
          type: integer
          format: int64
          example: 1
        nombre:
          type: string
          minLength: 2
          example: Festival de Primavera
        descripcion:
          type: string
          example: Festival de música en directo
        fechaEvento:
          type: string
          format: date
          example: '2026-06-20'
        horaEvento:
          type: string
          example: '21:00:00'
        precioEntrada:
          type: number
          format: float
          minimum: 0
          example: 45.5
        aforoMaximo:
          type: integer
          minimum: 1
          example: 12000
        entradasDisponibles:
          type: integer
          minimum: 0
          example: 8500
        cancelado:
          type: boolean
          example: false
        presencial:
          type: boolean
          example: true
        categoria:
          type: string
          example: Festival
        usuario:
          $ref: '#/components/schemas/UsuarioRef'
        artista:
          $ref: '#/components/schemas/ArtistaRef'
        recinto:
          $ref: '#/components/schemas/RecintoRef'
      required:
        - nombre
        - fechaEvento
        - horaEvento
        - usuario
        - artista
        - recinto
...